+++ record/CHANGELOG.md
@@ -3,6 +3,8 @@
 * fix: recorder state streams.
 * core: fmedia updated to 1.28 (linux / windows).
 * fix: Add error details on record start (ios).
+- feat: Add input devices listing.
+- feat: Add number of channels.
 
 ## 4.2.0-dev.1
 * feat: Add recorder state streams.