+++ record_macos/CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.1.0
+## 0.2.0
+- feat: Add input devices listing.
+- feat: Add number of channels.
 
+## 0.1.0
 * core: Initial release.